Moving files so platformio linker works
