+++ include/RwStatus.hpp
@@ -32,7 +32,7 @@ class RwStatus {
 };
 }  // namespace rw_status
 
-/* 
+/*
 These define's must be placed at the beginning before #include "TimerInterrupt.h"
 _TIMERINTERRUPT_LOGLEVEL_ from 0 to 4
 Don't define _TIMERINTERRUPT_LOGLEVEL_ > 0. Only for special ISR debugging only. Can hang the system.
@@ -62,12 +62,34 @@ volatile float wheel_rpm[rw_status::kNumWheels] = {0, 0, 0, 0};     // index to 
 volatile uint32_t global_time;                     // 0.1 ms
 // interrupt functions to calculate the rpm of each motor
 
+// TODO: there is an issue where it seems that placing these in RwStatus.cpp can't be handled by platformio.
+// Look into that
+
 // global timer used for rpm calculation
-void TimerHandler();
+void TimerHandler() {
+  global_time++;
+}
+
 // these must have "void FunctionName()" signature
-void ReadRpm0();
-void ReadRpm1();
-void ReadRpm2();
-void ReadRpm3();
+void ReadRpm0() {
+  wheel_rpm[0] = (kSecPerMin*kRotPerFG*kGlobalRate
+    /(float) (global_time - prev_time[0])) - kZeroRpm;
+  prev_time[0] = global_time;
+}
+void ReadRpm1() {
+  wheel_rpm[1] = (kSecPerMin*kRotPerFG*kGlobalRate
+    /(float) (global_time - prev_time[1])) - kZeroRpm;
+  prev_time[1] = global_time;
+}
+void ReadRpm2() {
+  wheel_rpm[2] = (kSecPerMin*kRotPerFG*kGlobalRate
+    /(float) (global_time - prev_time[2])) - kZeroRpm;
+  prev_time[2] = global_time;
+}
+void ReadRpm3() {
+  wheel_rpm[3] = (kSecPerMin*kRotPerFG*kGlobalRate
+    /(float) (global_time - prev_time[3])) - kZeroRpm;
+  prev_time[3] = global_time;
+}
 }  // namespace interrupt
 #endif  // RW_SRC_RWSTATUSCHECKER_HPP_
