+++ include/utility/quaternion.hpp
