+++ platformio.ini
@@ -12,7 +12,7 @@
 platform = atmelavr
 board = megaatmega2560
 framework = arduino
-lib_deps = 
+lib_deps =
 	khoih-prog/TimerInterrupt@^1.8.0
 	arduino-libraries/SD@^1.2.4
 	adafruit/Adafruit BNO08x@^1.2.3
